refactor: 컴포넌트 분리 - 복잡도를 낮추기 위해 장바구니 컴포넌트 분리
